+++ Combine_SwiftUI/Combine_SwiftUI/ShoppingCart/ShoppingCartModel.swift
@@ -1,0 +1,12 @@
+//
+//  ShoppingCartModel.swift
+//  Combine_SwiftUI
+//
+
+import Foundation
+
+struct Foodstuffs {
+    var name: String
+    var cost: Int
+    var id: Int
+}
